Revert "feat: call configResolved hook on new plugins"
This reverts commit 0e552f5514e9ea4e39f3e54577c21c25a89c96b4.
