+++ packages/vite/src/node/build.ts
@@ -49,11 +49,7 @@ import {
   normalizePath,
   partialEncodeURIPath,
 } from './utils'
-import {
-  getHookHandler,
-  perEnvironmentPlugin,
-  resolveEnvironmentPlugins,
-} from './plugin'
+import { perEnvironmentPlugin, resolveEnvironmentPlugins } from './plugin'
 import { manifestPlugin } from './plugins/manifest'
 import type { Logger } from './logger'
 import { dataURIPlugin } from './plugins/dataUri'
@@ -70,6 +66,7 @@ import {
 import { completeSystemWrapPlugin } from './plugins/completeSystemWrap'
 import { mergeConfig } from './publicUtils'
 import { webWorkerPostPlugin } from './plugins/worker'
+import { getHookHandler } from './plugins'
 import {
   BaseEnvironment,
   getDefaultResolvedEnvironmentOptions,
