+++ packages/vite/src/node/config.ts
@@ -29,7 +29,6 @@ import type {
   PluginOption,
   PluginWithRequiredHook,
 } from './plugin'
-import { getHookHandler, getSortedPluginsByHook } from './plugin'
 import type {
   BuildEnvironmentOptions,
   BuilderOptions,
@@ -67,7 +66,12 @@ import {
   normalizeAlias,
   normalizePath,
 } from './utils'
-import { createPluginHookUtils, resolvePlugins } from './plugins'
+import {
+  createPluginHookUtils,
+  getHookHandler,
+  getSortedPluginsByHook,
+  resolvePlugins,
+} from './plugins'
 import type { ESBuildOptions } from './plugins/esbuild'
 import type {
   EnvironmentResolveOptions,
