+++ packages/vite/src/node/server/hmr.ts
@@ -7,7 +7,7 @@ import type { RollupError } from 'rollup'
 import { CLIENT_DIR } from '../constants'
 import { createDebugger, normalizePath } from '../utils'
 import type { InferCustomEventPayload, ViteDevServer } from '..'
-import { getHookHandler } from '../plugin'
+import { getHookHandler } from '../plugins'
 import { isCSSRequest } from '../plugins/css'
 import { isExplicitImportRequired } from '../plugins/importAnalysis'
 import { getEnvFilesForMode } from '../env'
