+++ packages/vite/src/node/server/pluginContainer.ts
@@ -77,8 +77,7 @@ import {
   timeFrom,
 } from '../utils'
 import { FS_PREFIX } from '../constants'
-import { createPluginHookUtils } from '../plugins'
-import { getHookHandler } from '../plugin'
+import { createPluginHookUtils, getHookHandler } from '../plugins'
 import { cleanUrl, unwrapId } from '../../shared/utils'
 import type { PluginHookUtils } from '../config'
 import type { Environment } from '../environment'
